mxruntime -> libinfo
