+++ python/mxnet/libinfo.py
@@ -21,6 +21,13 @@ from __future__ import absolute_import
 import os
 import platform
 import logging
+import ctypes
+import enum
+from .base import _LIB, check_call, mx_uint, py_str
+
+
+# current version
+__version__ = "1.5.0"
 
 
 def find_lib_path():
@@ -76,6 +83,7 @@ def find_lib_path():
         os.environ['PATH'] = os.environ['PATH'] + ';' + os.path.dirname(lib_path[0])
     return lib_path
 
+
 def find_include_path():
     """Find MXNet included header files.
 
@@ -110,5 +118,62 @@ def find_include_path():
                                ' or ' + src_incl_path + '\n')
 
 
-# current version
-__version__ = "1.5.0"
+def _feature_names_available():
+    """
+
+    :return:
+    """
+    feature_list = ctypes.POINTER(ctypes.c_char_p)()
+    feature_list_sz = ctypes.c_size_t()
+    check_call(_LIB.MXRuntimeFeatureList(ctypes.byref(feature_list_sz), ctypes.byref(feature_list)))
+    feature_names = []
+    for i in range(feature_list_sz.value):
+        feature_names.append(py_str(feature_list[i]))
+    return feature_names
+
+
+Feature = enum.Enum('Feature', {name: index for index, name in enumerate(_feature_names_available())})
+
+
+def features_available():
+    """
+    Returns
+    -------
+    features: list of Feature enum
+        Features available in the backend which includes disabled and enabled ones
+    """
+    return list(Feature)
+
+
+def has_feature_index(feature):
+    """
+    Check the library for compile-time feature at runtime
+
+    Parameters
+    ----------
+    feature : int
+        An integer representing the feature to check
+
+    Returns
+    -------
+    boolean
+        True if the feature is enabled, false otherwise
+    """
+    res = ctypes.c_bool()
+    check_call(_LIB.MXRuntimeHasFeature(mx_uint(feature), ctypes.byref(res)))
+    return res.value
+
+
+def features_enabled():
+    """
+    Returns
+    -------
+    features: list of Feature enum
+        list of enabled features in the back-end
+    """
+    res = []
+    for f in Feature:
+        if has_feature_index(f.value):
+            res.append(f)
+    return res
+
